pyproject.toml
[tool.poetry]
name = "PriviaStajApp"
version = "0.1.0"
description = ""
authors = ["KaderMiyanyedi"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1"
Pillow = "^7.2.0"
djangorestframework = "^3.11.1"
drf-yasg = "^1.17.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pip = "^20.2.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
